Sync with changes to pstack
To save memory the pstack library now treats DIEs returned to callers
as ephemeral objects. We can get a DIE back out of a unit be presenting
the unit with the offset of the DIE you want. DIEs also do not contain
pointers to their parent, so we can't implement the scoping operations
from C as easily . Instead modify the tree-traversing python code to
keep track of namespacepaths.

Also, while playing here, make the namespace lookup more useful:
"a" matches only objects in the global namespace. "a::b" matches b in
the a namespace, as you'd expect.