Make enable/disable on a mock safe.
When we "enable"/"disable" a mock, we previously always did the actions
required to enable and disable it, regardless of whether the mock was
already in the required state.

We now count the invocations of enable/disable, so we can make sure we
don't disable an already-disabled mock

This is important for the destructor, where we call disable() to make
sure the mock is disabled. This is useful when someone forgets to disable
a mock, but if it always acts, then a random GC can impact an existing
live mock for the same function

Refactor out the common bit of having the python object head and the
"disable" count into a base class for all mock objects.

While here, also fix up the member access in the mock structures, and
remove the "function" `const char *` field, which is not valid after
the constructor exits.

Change-Id: Ieb8d320a9d0c3b4879e9f9a15fdec7856a724182