Fix finding C++ methods
C++ methods are generally declared with a "DW_AT_declaration" inside the
class (giving their args and return type), and then when defined, the
definition refers back to this with a DW_AT_specification.

We were previously trying to use the declaration, but that slows things
down massively if we look at all function declarations, as we try and
search for the definition out-of-band of the normal linear walk.
Instead, just wait until we get the function definition, and use the
specification it refers to.