Deal with anonymous bitfields.
Bitfields in C can be anonymous - eg:

```
struct x {
   int field1 : 1;
   int : 20;
   int field2 : 16;
};
```

In the above structure, field2 needs to occupy a second "int" in the
structure.

The debug information includes two extra attributes for bitfields -
DW_AT_bit_size an DW_AT_bit_offset to represent where in the data object
they occupy their bits are. gcc at least does *not* output any
information at all about the anon bitfield - it merely adjust the bit
offset for the next named field appropriately.

In ctypes then, we must add a new explicit bitfield to take the place of
the anonymous one. We can detect when a new bitfield has an unexpected
offset, and insert the required padding before it.