CTypeGen: Flatten out all the namespace handling
This removes the complicated hierarchical namespace handling, and
instead uses the DIEs "fullname" globally. A DIEs fullname is the fully
qualified name of the DIE with all enclosing namespaces as a tuple.
Eg, `"Foo::Bar::Baz"` has a fullname of `("Foo", "Bar", "Baz")`

All names are generally now handled as tuples, and for most cases, for C
code, the tuples are single-valued, eg `("fork",)`

We also remove some other unnecessary complications, like having separate
sets for declared and defined types - types now have a "declared" and
"defined" fields, and we use a single container, "types", for them.

The only references to types, variables, and functions is now via the
so-named fields of the type resolver. They are indexed by the "fullname",
and, for types, we have a second level index of "tag", so we can now
keep track of types for typedefs, unions, structs, etc, separately.

(The second-level 'tag' index for types is to allow for the follow-on
mut that will allow us to have typedefs and structs/unions of the same
name to co-exist)

Without the namespaces, ths callback functions for testing if the caller
wants a type/function/variable installed now just take the DIE object,
and can grab the "fullname" if needed.

Provide efficient verison of makeFullname, now that we make more extensive
use of die.fullname(). This saves making C++ vectors before converting
to python tuples.

Change-Id: I5ea705ae3f2722c217f57ad0ac4154ca93e03a7d