Allow types filter to return hints as well as true/false
Previously the types filter allowed you to either specify a list of hint
objects for types you wanted rendered, or a predicate function that
returned true or false to indicate if you wanted a DIE rendered.

The latter case did not allow any hinting control over the rendered
types.

This change allows the filter function to return hints for the type. We
remove the name-centric treatment of PythonType, and maintain the set
of hints we need to require as a mapping from type ("Type") to hint
("PythonType"). Also, remove the link from the type itself to the hint,
and copy things as required, so we can potentially share hints across
types.