Do better with DW_AT_decl_file
When requesting DW_AT_decl_file attribute, return the string for the
filename, rather thn the index into the line table. This gives some
useful info for type filters (i.e., we can limit ourselves to only
finding definitions of types in certain headers.)

Also, catch exception for when parsing a macro fails, and allow for
unnamed types - this is useful particularly for enums.

Change-Id: Id1681d82703037e6da2f46831da95c840023eb43