Improve bitfield heuristics.
Deal with bitfields pain. gcc doesn't output anything for unnamed
bitfields, and we had problems with the existing heuristic if there were
empty bitfields at the end of a structure, or multiple anonymous
bitfields in the middle of one.

Fixing this requires moving the bitfield handling logic to the define()
function for structs, from where it was in findMembers. This makes more
sense anyway - findMembers is really just trying to work out what types
we need to define before we define the struct itself. We need to know
about the alignment of the fields to work out if there's missing padding
at the end of a struct.

We don't try to reconstitute the unnamed fields in that place, and just
use a char array to pad it out. (There's no point, because anything we
do with the bitfields will just be random guesses, and there's no way
to access them anyway)

Also, some changes to allow unaligned pointers on a per-struct-type basis,
rather than per field

Change-Id: If8f871433f3b316231dfe145966e30a5a04731ac