Deal with inheriting from empty types
C++ structures always have a size >= 1. However, if the structure is
used as a base class for another, then an empty struct will not
contribute anything to the size of its subclass. This is also true for
padding issues, but we are not dealing with that here - detecting this
common case is enough to let CTypeGen deal with most of the common
cases.

Also, deal with overrides for types on fields - if the die says a field
is a bitfield, but the override says otherwise, trust the override.

Change-Id: I2d546d44f93be304e201d8863d9315119f9d5247