CTypeGen: avoid "tail-padding" for empty types.
C++ treats empty class/struct types oddly. An empty type has size 1, but
if you inherit from it, you don't incur the penalty - this is the
"empty base optimization".

A recent change to CTypeGen had it pad structures that were mis-sized
because it can legitimately be required when you have anonymous bitfields,
etc, at the end of the struct. The empty-base-optimization was
triggering this, so explcitly avoid the padding for otherwise empty types

Change-Id: Iee093f550baad9ba83ce983488525a46998b4950