Don't look for "anonymous" types in other modules
For C, we assume that type names are generally "global" - if you have
a type,  and want to share the module it is in with other modules, we
will use existing types of the same name in explicitly provided modules
in place of locally defined ones (because each DSO must have its own
definition of each struct/class/enum etc that it uses).

This is useful when we have structure definitions in a "base" library
for example that get used elsewhere. We reduce the generated code size,
and allow the types in the different modules to appear to be the same,
and we don't need to cast from one definition of the type to another

If we have a type with no "name" attribute, it is "anonymous", and gets
a name based on the translation unit it is in, and the offset of the
DIE in that unit - eg "anon_struct_anon_1_6583_struct".  These may not
be unique across modules, though, and it makes no sense to apply the
rule that commonly named types are "the same" in that case. This avoids
random failures when the type collides with something from a provided
"base" library.