Add more missing types.
Apparently glibc on arora now has a "sizetype" primitive type, and
doesn't always encode a DW_AT_type for an enumeration.

Use c_ulong for sizetype - I've verified its encoding is signed, and
commensurate with c_ulong for 64- and 32- bit.

Just assume "c_uint" for enums with no DW_AT_type for now.

We should instead use the type encoding in DWARF for this instead.