Implement PRE mock thunks for aarch64.
PRE mocks allow you to register a python function that will be called
before a C one, when that C function is called via the GOT. The ARM
implementation was non-existent, so implement it.

Disclaimer: I have no idea what I am doing. I "learned" enough ARM asm
to do this, so it's probably not exactly pretty or idiomatic.