CTypeGen: use pc-relative addressing for x86-64
I originally copied the 32-bit version of this code for 64-bit x86_64.
However, all the call-next-instruction+pop-return-address machinations
are not required on x86_64 - we have proper PC-relative addressing. Use
that to simplify things a lot.

While here, clean up a bit and add comments. One functional change is to
pop all but the return address of the thunk stack before calling
function #2 - this allows deeper recursion for this function by using
far less space on the thunk stack.