CTypeGen: STOMP mocks for aarch64
Mostly for completeness, provide an implementation of STOMP mocks for
aarch64. These mocks are for when you don't have GOT indirection for
functions. As long as the function you want to mock is larger than some
minimum (20 bytes for aarch64), we can just patch the start of the
function with a branch to the mocking function.